chore: organize unit tests in a folder
