+++ .sfdevrc.json
@@ -1,10 +1,10 @@
 {
-  "test": {
-    "testsPath": "test/**/*.test.ts"
-  },
-  "wireit": {
     "test": {
-      "dependencies": ["test:compile", "test:only", "lint"]
+        "testsPath": "test/unit/**/*.test.ts"
+    },
+    "wireit": {
+        "test": {
+            "dependencies": ["test:compile", "test:only", "lint"]
+        }
     }
-  }
 }
