+++ .vscode/launch.json
@@ -1,79 +1,61 @@
 {
-  // Use IntelliSense to learn about possible attributes.
-  // Hover to view descriptions of existing attributes.
-  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "type": "node",
-      "request": "attach",
-      "name": "Attach",
-      "port": 9229,
-      "skipFiles": [
-        "<node_internals>/**"
-      ]
-    },
-    {
-      "name": "Run All Tests",
-      "type": "node",
-      "request": "launch",
-      "program": "${workspaceFolder}/node_modules/mocha/bin/mocha",
-      "args": [
-        "--inspect",
-        "--colors",
-        "test/**/*.test.ts"
-      ],
-      "env": {
-        "NODE_ENV": "development",
-        "SFDX_ENV": "development"
-      },
-      "sourceMaps": true,
-      "smartStep": true,
-      "internalConsoleOptions": "openOnSessionStart",
-      "preLaunchTask": "Compile tests"
-    },
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Run Current Test",
-      "program": "${workspaceFolder}/node_modules/mocha/bin/mocha",
-      "args": [
-        "--inspect",
-        "--colors",
-        "${file}"
-      ],
-      "env": {
-        "NODE_ENV": "development",
-        "SFDX_ENV": "development"
-      },
-      "sourceMaps": true,
-      "smartStep": true,
-      "internalConsoleOptions": "openOnSessionStart",
-      "preLaunchTask": "Compile tests"
-    },
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Setup - iOS",
-      "console": "integratedTerminal",
-      "program": "${workspaceFolder}/bin/run",
-      "args": [
-        "force:lightning:local:setup",
-        "-p",
-        "ios"
-      ]
-    },
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "Setup - Android",
-      "console": "integratedTerminal",
-      "program": "${workspaceFolder}/bin/run",
-      "args": [
-        "force:lightning:local:setup",
-        "-p",
-        "android"
-      ]
-    },
-  ]
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "type": "node",
+            "request": "attach",
+            "name": "Attach",
+            "port": 9229,
+            "skipFiles": ["<node_internals>/**"]
+        },
+        {
+            "name": "Run All Unit Tests",
+            "type": "node",
+            "request": "launch",
+            "program": "${workspaceFolder}/node_modules/mocha/bin/mocha",
+            "args": ["--inspect", "--colors", "test/unit/**/*.test.ts"],
+            "env": {
+                "NODE_ENV": "development",
+                "SFDX_ENV": "development"
+            },
+            "sourceMaps": true,
+            "smartStep": true,
+            "internalConsoleOptions": "openOnSessionStart",
+            "preLaunchTask": "Compile tests"
+        },
+        {
+            "type": "node",
+            "request": "launch",
+            "name": "Run Current Test",
+            "program": "${workspaceFolder}/node_modules/mocha/bin/mocha",
+            "args": ["--inspect", "--colors", "${file}"],
+            "env": {
+                "NODE_ENV": "development",
+                "SFDX_ENV": "development"
+            },
+            "sourceMaps": true,
+            "smartStep": true,
+            "internalConsoleOptions": "openOnSessionStart",
+            "preLaunchTask": "Compile tests"
+        },
+        {
+            "type": "node",
+            "request": "launch",
+            "name": "Setup - iOS",
+            "console": "integratedTerminal",
+            "program": "${workspaceFolder}/bin/run",
+            "args": ["force:lightning:local:setup", "-p", "ios"]
+        },
+        {
+            "type": "node",
+            "request": "launch",
+            "name": "Setup - Android",
+            "console": "integratedTerminal",
+            "program": "${workspaceFolder}/bin/run",
+            "args": ["force:lightning:local:setup", "-p", "android"]
+        }
+    ]
 }
