+++ test/unit/common/AndroidTypes.test.ts
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
 import { expect } from 'chai';
-import { AndroidPackages } from '../../src/common/AndroidTypes.js';
+import { AndroidPackages } from '../../../src/common/AndroidTypes.js';
 import { AndroidMockData } from './AndroidMockData.js';
 
 describe('Android types tests', () => {
