+++ test/unit/common/CommonUtils.test.ts
@@ -11,7 +11,7 @@ import { Messages } from '@salesforce/core';
 import { TestContext } from '@salesforce/core/testSetup';
 import { stubMethod } from '@salesforce/ts-sinon';
 import { expect } from 'chai';
-import { CommonUtils } from '../../src/common/CommonUtils.js';
+import { CommonUtils } from '../../../src/common/CommonUtils.js';
 
 Messages.importMessagesDirectoryFromMetaUrl(import.meta.url);
