+++ test/unit/common/testGetAppBundleScript
